:recycle: Refactored auth-service and echo-service into helm charts
The previous Kubernetes manifests for auth-service and echo-service were deleted and replaced with new helm charts for better modularity and scalability. The chart includes service and deployment YAML files for each service, and a values.yaml file for defining variables. Additionally, an EnvoyFilter for JWT authorization was added using helm chart.