:sparkles: Add VirtualService configuration for project
A new Istio VirtualService configuration has been added for the project. This configuration defines custom routing rules in the `virtualservice.yaml` file. The rules specify that all incoming traffic should be directed to either the auth service or the echo service based on the request URI prefix. This routing is necessary to ensure correct interactions between